Refactor iOS webviews demos: - Added new assembly files. - Introduced load_webview.r2 for 0096 as well - Updated output.txt files - Modified run.sh to streamline execution with the new load_webview.r2 configuration. - Update demo content to reflect all these changes including detailed disassembly analysis
