+++ demos/ios/MASVS-PLATFORM/MASTG-DEMO-0096/run.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-r2 -q -i ../MASTG-DEMO-0095/load_webview.r2 -A ../MASTG-DEMO-0095/MASTestApp > output.txt
+r2 -q -i load_webview.r2 -A ../MASTG-DEMO-0095/MASTestApp > output.txt
